requirements.txt
streamlit
torch
torchvision
Pillow
huggingface_hub

